requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.42.0
django-common-helpers==0.9.2
django-crispy-forms==1.9.1
django-cron==0.5.1
django-crontab==0.7.1
django-cropperjs==0.0.5
django-dbbackup==3.3.0
django-mysql==3.7.1
django-pwa==1.0.9
django-session-timeout==0.1.0
djangorestframework==3.11.0
docopt==0.6.2
Faker==4.1.1
gunicorn==20.0.4
html5lib==1.1
idna==2.10
mutagen==1.44.0
num2words==0.5.10
oauthlib==3.1.0
pdfkit==0.6.1
Pillow==7.1.2
git+git://github.com/ojii/pymaging-png.git#egg=pymaging-png
git+git://github.com/ojii/pymaging.git#egg=pymaging
PyMySQL==0.9.3
PyPDF2==1.26.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
qrcode==6.1
reportlab==3.5.44
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
webencodings==0.5.1
xhtml2pdf==0.2.4
